Move glue logs logic to logs submodule
